reorganize repository to make structure clearer and cleaner
