+++ README.md
@@ -7,14 +7,37 @@ https://iabtechlab.com/standards/artf/
 
 #### How to get started
 
-Download the openRTB official 2.6 Protocol Buffers specification from https://github.com/InteractiveAdvertisingBureau/openrtb2.x/blob/main/proto/src/main/com/iabtechlab/openrtb/v2/openrtb.proto to this directory.
+`make generate` will fetch the OpenRTB 2.6 proto automatically. If you want to refresh it manually, run `make fetch-openrtb`.
 
 From the command line:
 
 1. Install `make` and the latest version of `protoc`.
-2. Open the `Makefile` and choose the language(s) for which the Protocol Buffers
-   object code should be generated.
-3. Run `make`.
+2. Choose the language(s) for which the Protocol Buffers object code should be generated.
+3. Run `make generate` or `make generate LANG=java`.
+
+#### Package names in generated code
+
+If you want generated code to use your own repo/package name, set the package options in the `.proto` files before generating code.
+
+Examples:
+
+- Go: `option go_package = "github.com/your-org/your-repo/pkg/pb/artf;artf";`
+- Java: `option java_package = "com.yourorg.yourrepo.artf";`
+- C++: `option cpp_namespace = "your::org::yourrepo::artf";`
+- Rust: use the generator/tooling-specific crate/module naming options for your Rust codegen pipeline
+
+#### Language selection with `make generate`
+
+You can pick one or more languages by passing a variable to `make`:
+
+```bash
+make generate
+make generate LANG=java
+make generate LANG=go,cpp,java
+make generate LANGS=rust
+```
+
+`LANG` and `LANGS` are both accepted; if neither is set, Go is generated by default.
 
 #### Contact
 For more information, or to get involved, please email support@iabtechlab.com.
@@ -68,7 +91,7 @@ The following tools must be installed to generate protobuf code:
 export PATH="$PATH:$(go env GOPATH)/bin"
 ```
 
-Go module dependencies (managed via `go.mod`):
+Go module dependencies are managed in the workspace modules:
 
 | Package | Version | Purpose |
 |---------|---------|---------|
@@ -79,55 +102,56 @@ Go module dependencies (managed via `go.mod`):
 #### Build and Run
 
 ```bash
-# Install dependencies
-make deps
-
-# Generate protobuf code
+# Generate protobuf code (root repo task)
 make generate
 
-# Build the server
-make build
+# Build the service binary
+make -C examples/golang build
 
-# Run with all services enabled
-make run-all
+# Run with all interfaces enabled
+make -C examples/golang run-all
 
-# Run in development mode (verbose)
-make run-dev
+# Run specific service modes
+make -C examples/golang run-grpc
+make -C examples/golang run-mcp
+make -C examples/golang run-web
 ```
 
 #### Docker Deployment
 
 ```bash
-# Build Docker image
-make docker-build
+# Build the service Docker image
+make -C examples/golang docker-build
 
-# Run with Docker
-make docker-run-all
+# Run the image directly
+make -C examples/golang docker-run-all
 
-# Or use docker-compose
-make docker-compose-up
+# Or use the service compose file
+cd examples/golang && docker compose up --build
 ```
 
 ### Architecture
 
 ```
 .
-├── cmd/agent/           # Main agent entry point
-├── internal/
-│   ├── agent/           # gRPC agent implementation
-│   ├── handlers/        # Mutation handlers for different intents
-│   ├── health/          # Kubernetes health check endpoints
-│   ├── mcp/             # MCP server implementation
-│   └── web/             # Web UI for testing
-├── pkg/pb/              # Generated protobuf Go code
+├── examples/golang/
+│   ├── cmd/agent/       # Main agent entry point
+│   └── internal/
+│       ├── agent/       # gRPC agent implementation
+│       ├── handlers/    # Mutation handlers for different intents
+│       ├── health/      # Kubernetes health check endpoints
+│       ├── mcp/         # MCP server implementation
+│       └── web/         # Web UI for testing
+├── pkg/                  # Shared Go module for generated protobuf code
+│   └── pb/              # Generated protobuf Go code
 ├── proto/               # Protocol buffer definitions
 │   ├── agenticrtbframework.proto  # ARTF service definition
 │   └── com/iabtechlab/openrtb/    # OpenRTB v2.6 definitions
 ├── samples/             # Sample ORTB payloads for testing
 ├── docs/                # Specifications and documentation
 ├── scripts/             # Build and utility scripts
-├── Dockerfile           # Container build definition
-└── docker-compose.yml   # Local development setup
+├── Makefile             # Root proto generation tasks
+└── go.work              # Go workspace for repo modules
 ```
 
 ### API
