+++ examples/golang/federation.example.yaml
