+++ examples/golang/go.mod
@@ -1,13 +1,16 @@
-module github.com/iabtechlab/agentic-rtb-framework
+module github.com/iabtechlab/agentic-rtb-framework/examples/golang
 
 go 1.23.0
 
 require (
+	github.com/iabtechlab/agentic-rtb-framework/pkg v0.0.0
 	github.com/mark3labs/mcp-go v0.43.1
 	google.golang.org/grpc v1.64.0
 	google.golang.org/protobuf v1.34.1
 	gopkg.in/yaml.v3 v3.0.1
 )
+
+replace github.com/iabtechlab/agentic-rtb-framework/pkg => ../../pkg
 
 require (
 	github.com/bahlo/generic-list-go v0.2.0 // indirect
