+++ examples/golang/internal/federation/client.go
