+++ examples/golang/internal/federation/config.go
