+++ examples/golang/internal/federation/manager.go
