+++ examples/golang/internal/health/health.go
