+++ examples/golang/internal/web/handler.go
