+++ examples/golang/internal/web/static/.gitkeep
