+++ examples/golang/internal/web/static/container.html
