+++ examples/golang/internal/web/static/spec.html
