+++ examples/golang/internal/web/templates/index.html
