+++ examples/rust/Dockerfile
