+++ examples/rust/build.rs
