+++ examples/rust/cargo.toml
