+++ examples/rust/proto/com/iabtechlab/openrtb/v2/openrtb.proto
