+++ examples/rust/src/bidder/evaluator.rs
