+++ examples/rust/src/bidder/mod.rs
