+++ examples/rust/src/main.rs
