+++ examples/rust/src/mutation/types.rs
