+++ examples/rust/src/proto/mod.rs
