+++ examples/rust/src/service/grpc.rs
