+++ examples/rust/src/service/http.rs
