+++ go.work
@@ -1,0 +1,6 @@
+go 1.23.0
+
+use (
+	./pkg
+	./examples/golang
+)
