+++ proto/agenticrtbframework.proto
@@ -3,7 +3,7 @@ edition = "2023";
 package com.iabtechlab.bidstream.mutation.v1;
 
 // Import OpenRTB definitions for BidRequest and BidResponse
-import "com/iabtechlab/openrtb/v2.6/openrtb.proto";
+import "com/iabtechlab/openrtb/v2/openrtb.proto";
 
 // ------------------- Messages -------------------
