+++ scripts/generate.sh
@@ -1,49 +1,166 @@
-#!/bin/bash
-# Generate Go code from protobuf definitions
-# Requires: protoc, protoc-gen-go, protoc-gen-go-grpc
+#!/usr/bin/env bash
+# Generate protobuf bindings for one or more target languages.
 #
-# OpenRTB 2.6 proto is fetched from IAB Tech Lab repository:
-# https://github.com/IABTechLab/openrtb-proto-v2
+# Examples:
+#   ./scripts/generate.sh --lang go
+#   ./scripts/generate.sh --lang go,rust
+#   ./scripts/generate.sh --lang all
 
-set -e
+set -euo pipefail
 
 SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 PROJECT_ROOT="$(dirname "$SCRIPT_DIR")"
 PROTO_DIR="$PROJECT_ROOT/proto"
-OUT_DIR="$PROJECT_ROOT/pkg/pb"
+GO_OPENRTB_MODULE="github.com/iabtechlab/agentic-rtb-framework/pkg/pb/openrtb"
+GO_ARTF_MODULE="github.com/iabtechlab/agentic-rtb-framework/pkg/pb/artf"
+SUPPORTED_LANGS=(go rust cpp java)
+PROTO_FILES=()
+while IFS= read -r proto_file; do
+  PROTO_FILES+=("$proto_file")
+done < <(find "$PROTO_DIR" -type f -name '*.proto' | sort)
 
-# OpenRTB proto location (downloaded by make fetch-openrtb)
-OPENRTB_PROTO="$PROTO_DIR/com/iabtechlab/openrtb/v2/openrtb.proto"
-
-# Check if OpenRTB proto exists
-if [ ! -f "$OPENRTB_PROTO" ]; then
-  echo "Error: OpenRTB proto not found at $OPENRTB_PROTO"
-  echo "Run 'make fetch-openrtb' to download it from IAB Tech Lab repository"
+if [ "${#PROTO_FILES[@]}" -eq 0 ]; then
+  echo "Error: no .proto files found under $PROTO_DIR"
   exit 1
 fi
 
-# Ensure output directories exist
-mkdir -p "$OUT_DIR/openrtb"
-mkdir -p "$OUT_DIR/artf"
+usage() {
+  cat <<'EOF'
+Usage: ./scripts/generate.sh [--lang LANG[,LANG...]]
 
-echo "Generating Go code from protobuf definitions..."
+Supported languages: go, rust, cpp, java, all
 
-# Generate OpenRTB 2.6 types (from IAB Tech Lab openrtb-proto-v2)
-echo "  - OpenRTB 2.6 types..."
-protoc \
-  --proto_path="$PROTO_DIR" \
-  --go_out="$PROJECT_ROOT" \
-  --go_opt=module=github.com/iabtechlab/agentic-rtb-framework \
-  "$OPENRTB_PROTO"
+Examples:
+  ./scripts/generate.sh --lang go
+  ./scripts/generate.sh --lang rust
+  ./scripts/generate.sh --lang cpp
+  ./scripts/generate.sh --lang java
+  ./scripts/generate.sh --lang go,rust
+  ./scripts/generate.sh --lang all
+EOF
+}
 
-# Generate ARTF service and types
-echo "  - ARTF service and types..."
-protoc \
-  --proto_path="$PROTO_DIR" \
-  --go_out="$PROJECT_ROOT" \
-  --go_opt=module=github.com/iabtechlab/agentic-rtb-framework \
-  --go-grpc_out="$PROJECT_ROOT" \
-  --go-grpc_opt=module=github.com/iabtechlab/agentic-rtb-framework \
-  "$PROTO_DIR/agenticrtbframework.proto"
+if ! command -v protoc >/dev/null 2>&1; then
+  echo "Error: protoc is required but not installed or not on PATH."
+  exit 1
+fi
 
-echo "Done! Generated files in $OUT_DIR"
+LANGS=()
+while [ "$#" -gt 0 ]; do
+  case "$1" in
+   -l|--lang|--language)
+     shift
+     if [ "$#" -eq 0 ]; then
+       echo "Error: --lang requires a value."
+       usage
+       exit 1
+     fi
+     IFS=',' read -r -a EXTRA_LANGS <<< "$1"
+     for lang in "${EXTRA_LANGS[@]}"; do
+       LANGS+=("${lang//[[:space:]]/}")
+     done
+     ;;
+   -h|--help)
+     usage
+     exit 0
+     ;;
+   *)
+     echo "Error: unknown argument: $1"
+     usage
+     exit 1
+     ;;
+  esac
+  shift
+done
+
+if [ "${#LANGS[@]}" -eq 0 ]; then
+  LANGS=("go")
+fi
+
+get_output_dir() {
+  local lang="$1"
+  local target
+  local idx
+
+  for idx in "${!SUPPORTED_LANGS[@]}"; do
+   target="${SUPPORTED_LANGS[$idx]}"
+   if [ "$target" = "$lang" ]; then
+     echo "$PROJECT_ROOT/pkg"
+     return 0
+   fi
+  done
+
+  echo "Error: unsupported language '$lang'" >&2
+  exit 1
+}
+
+require_tool() {
+  local tool="$1"
+  local purpose="$2"
+  if ! command -v "$tool" >/dev/null 2>&1; then
+   echo "Error: $tool is required for $purpose generation but is not installed on PATH."
+   exit 1
+  fi
+}
+
+generate_language() {
+  local lang
+  lang="$(printf '%s' "$1" | tr '[:upper:]' '[:lower:]')"
+  local out_dir
+  out_dir="$(get_output_dir "$lang")"
+  local -a args
+  local normalized_flag
+  local lang_upper
+
+  lang_upper="$(printf '%s' "$lang" | tr '[:lower:]' '[:upper:]')"
+  args=(--proto_path="$PROTO_DIR")
+
+  if [ "$lang" = "go" ]; then
+   require_tool protoc-gen-go "Go"
+   require_tool protoc-gen-go-grpc "Go"
+   args+=(--go_out="$out_dir" --go_opt=module="github.com/iabtechlab/agentic-rtb-framework/pkg")
+   args+=(--go_opt="Magenticrtbframework.proto=$GO_ARTF_MODULE")
+   args+=(--go_opt="Magenticrtbframeworkservices.proto=$GO_ARTF_MODULE")
+   args+=(--go_opt="Mcom/iabtechlab/openrtb/v2/openrtb.proto=$GO_OPENRTB_MODULE")
+   args+=(--go-grpc_out="$out_dir" --go-grpc_opt=module="github.com/iabtechlab/agentic-rtb-framework/pkg")
+   args+=(--go-grpc_opt="Magenticrtbframework.proto=$GO_ARTF_MODULE")
+   args+=(--go-grpc_opt="Magenticrtbframeworkservices.proto=$GO_ARTF_MODULE")
+   args+=(--go-grpc_opt="Mcom/iabtechlab/openrtb/v2/openrtb.proto=$GO_OPENRTB_MODULE")
+   echo "Generating Go bindings into $out_dir..."
+   protoc "${args[@]}" "${PROTO_FILES[@]}"
+   echo "Done! Generated Go files in $out_dir"
+   return 0
+  fi
+
+  if [ "$lang" = "rust" ]; then
+   if ! command -v protoc-gen-rust >/dev/null 2>&1 && ! command -v protoc-gen-tonic >/dev/null 2>&1; then
+     echo "Error: rust generation requires protoc-gen-rust and/or protoc-gen-tonic in PATH."
+     exit 1
+   fi
+   mkdir -p "$out_dir"
+   args+=(--rust_out="$out_dir")
+   if command -v protoc-gen-tonic >/dev/null 2>&1; then
+     args+=(--tonic_out="$out_dir")
+   fi
+   echo "Generating Rust bindings into $out_dir..."
+   protoc "${args[@]}" "${PROTO_FILES[@]}"
+   echo "Done! Generated Rust files in $out_dir"
+   return 0
+  fi
+
+  mkdir -p "$out_dir"
+  echo "Generating ${lang_upper} bindings into $out_dir..."
+  protoc "${args[@]}" "--${lang}_out=$out_dir" "${PROTO_FILES[@]}"
+  echo "Done! Generated ${lang_upper} files in $out_dir"
+}
+
+for lang in "${LANGS[@]}"; do
+  normalized="$(printf '%s' "$lang" | tr '[:upper:]' '[:lower:]')"
+  if [ "$normalized" = "all" ]; then
+   for target in "${SUPPORTED_LANGS[@]}"; do
+     generate_language "$target"
+   done
+  else
+   generate_language "$normalized"
+  fi
+done
